move init integration tests to unit tests
