+++ pkg/skaffold/initializer/init/.gitignore
@@ -1,0 +1,1 @@
+skaffold.yaml.out
