+++ pkg/skaffold/initializer/init/hello/k8s-pod.yaml
