+++ pkg/skaffold/initializer/init/hello/skaffold.yaml
