+++ pkg/skaffold/initializer/init/ignore-tags/Dockerfile
