+++ pkg/skaffold/initializer/init/ignore-tags/k8s-pod.yaml
