+++ pkg/skaffold/initializer/init/microservices/leeroy-app/kubernetes/deployment.yaml
