+++ pkg/skaffold/initializer/init/microservices/leeroy-web/Dockerfile
