+++ pkg/skaffold/initializer/init/microservices/leeroy-web/kubernetes/deployment.yaml
