+++ pkg/skaffold/initializer/init/microservices/skaffold.yaml
